+++ test/test.js
@@ -13,16 +13,19 @@ function evaluate( bundle ) {
 	return module.exports;
 }
 
+// Short-hand for rollup using the typescript plugin.
+function bundle( main ) {
+	return rollup.rollup({
+		entry: main,
+		plugins: [ typescript() ]
+	});
+}
+
 describe( 'rollup-plugin-typescript', function () {
 	this.timeout( 5000 );
 
 	it( 'runs code through typescript', function () {
-		return rollup.rollup({
-			entry: 'sample/basic/main.ts',
-			plugins: [
-				typescript()
-			]
-		}).then( function ( bundle ) {
+		return bundle( 'sample/basic/main.ts' ).then( function ( bundle ) {
 			const code = bundle.generate().code;
 
 			assert.ok( code.indexOf( 'number' ) === -1, code );
@@ -30,13 +33,20 @@ describe( 'rollup-plugin-typescript', function () {
 		});
 	});
 
+	it( 'does not duplicate helpers', function () {
+		return bundle( 'sample/dedup-helpers/main.ts' ).then( function ( bundle ) {
+			const code = bundle.generate().code;
+
+			// The `__extends` function is defined in the bundle.
+			assert.ok( code.indexOf( 'function __extends' ) > -1, code );
+
+			// No duplicate `__extends` helper is defined.
+			assert.equal( code.indexOf( '__extends$1' ), -1, code );
+		});
+	});
+
 	it( 'transpiles `export class A` correctly', function () {
-		return rollup.rollup({
-			entry: 'sample/export-class-fix/main.ts',
-			plugins: [
-				typescript()
-			]
-		}).then( function ( bundle ) {
+		return bundle( 'sample/export-class-fix/main.ts' ).then( function ( bundle ) {
 			const code = bundle.generate().code;
 
 			assert.equal( code.indexOf( 'class' ), -1, code );
@@ -47,12 +57,7 @@ describe( 'rollup-plugin-typescript', function () {
 	});
 
 	it( 'transpiles ES6 features to ES5 with source maps', function () {
-		return rollup.rollup({
-			entry: 'sample/import-class/main.ts',
-			plugins: [
-				typescript()
-			]
-		}).then( function ( bundle ) {
+		return bundle( 'sample/import-class/main.ts' ).then( function ( bundle ) {
 			const code = bundle.generate().code;
 
 			assert.equal( code.indexOf( 'class' ), -1, code );
@@ -62,23 +67,13 @@ describe( 'rollup-plugin-typescript', function () {
 	});
 
 	it( 'reports diagnostics and throws if errors occur during transpilation', function () {
-		return rollup.rollup({
-			entry: 'sample/syntax-error/missing-type.ts',
-			plugins: [
-				typescript()
-			]
-		}).catch( function ( error ) {
+		return bundle( 'sample/syntax-error/missing-type.ts' ).catch( function ( error ) {
 			assert.ok( error.message.indexOf( 'There were TypeScript errors' ) === 0, 'Should reject erroneous code.' );
 		});
 	});
 
 	it( 'should use named exports for classes', function () {
-		return rollup.rollup({
-			entry: 'sample/export-class/main.ts',
-			plugins: [
-				typescript()
-			]
-		}).then( function ( bundle ) {
+		return bundle( 'sample/export-class/main.ts' ).then( function ( bundle ) {
 			assert.equal( evaluate( bundle ).foo, 'bar' );
 		});
 	});
@@ -107,12 +102,7 @@ describe( 'rollup-plugin-typescript', function () {
 	});
 
 	it( 'should not resolve .d.ts files', function () {
-		return rollup.rollup({
-			entry: 'sample/dts/main.ts',
-			plugins: [
-				typescript()
-			]
-		}).then( function ( bundle ) {
+		return bundle( 'sample/dts/main.ts' ).then( function ( bundle ) {
 			assert.deepEqual( bundle.imports, [ 'an-import' ] );
 		});
 	});