+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # rollup-plugin-typescript changelog
 
 ## master
+* Do not duplicate TypeScript's helpers ([#24](https://github.com/rollup/rollup-plugin-typescript/issues/24)
 
 ### 0.4.1
 * Does not attempt resolve or transform `.d.ts` files ([#22](https://github.com/rollup/rollup-plugin-typescript/pull/22))